backend/requirements.txt
# Core Flask dependencies
Flask==2.3.3
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-mail==0.9.1

# Database
pymongo==4.5.0

# AI and ML dependencies
langchain==0.1.0
langchain-google-genai==1.0.1
langchain-community==0.0.13
faiss-cpu
textblob

# PDF processing
PyPDF2==3.0.1
reportlab==4.0.4

# Authentication
PyJWT==2.8.0

# Environment and utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Production server
gunicorn==21.2.0

# Development dependencies (optional)
pytest==7.4.2
pytest-flask==1.2.0
